requirements.txt
astroid==1.4.8
autopep8==1.3.2
backports.functools-lru-cache==1.3
behave==1.2.5
behave-django==0.5.0
Django==1.11
djangorestframework==3.6.2
dodgy==0.1.9
factory-boy==2.9.2
Faker==0.8.6
isort==4.2.5
lazy-object-proxy==1.2.2
parse==1.8.2
parse-type==0.4.2
pep8==1.7.0
pep8-naming==0.4.1
prospector==0.12.4
psycopg2==2.7.3.1
pycodestyle==2.3.1
pydocstyle==1.0.0
pylint==1.6.4
pylint-celery==0.3
pylint-common==0.2.2
pylint-django==0.7.2
pylint-flask==0.5
pylint-plugin-utils==0.2.4
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requirements-detector==0.5.2
setoptconf==0.2.0
six==1.11.0
text-unidecode==1.0
wrapt==1.10.8
